fix: holzman-rust compliance — safe indexing, checked conversions, clippy fixes
- gates.rs: replace unsafe indexing with .get()/ .get_mut() in DFS
- arena.rs: replace as casts and direct indexing with checked conversions
- exit_code.rs: add #[allow(clippy::as_conversions)] for intentional casts
- vb_cli: fix module declarations, imports, and macro references
- naming_scan: fix module structure and types
- compile: fix mod_compile_lowering part_02
- Remove empty test file vb_fzgdn_proptest.rs

+++ crates/vb_compile/src/mod_compile_lowering/part_02.rs
@@ -179,8 +179,8 @@ pub(super) fn lower_canonical_for_each(
             limit: usize::from(u16::MAX),
         }])
     })?;
-    let next_offset = u16::from(1u16.saturating_add(body_width_val));
-    let done_offset = u16::from(2u16.saturating_add(body_width_val));
+    let next_offset = 1u16.saturating_add(body_width_val);
+    let done_offset = 2u16.saturating_add(body_width_val);
     let next_step =
         checked_step_offset(id, next_offset, "for_each", "next").map_err(|e| CompileErrors(vec![e]))?;
     let done =
